# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PROfashion-Test API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/comments": {
            "get": {
                "tags": [
                    "Комментарии"
                ],
                "summary": "Получить список комментариев",
                "operationId": "0406f270b674fae17bf23b4206a3004f",
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            },
            "post": {
                "tags": [
                    "Комментарии"
                ],
                "summary": "Создать новый комментарий",
                "operationId": "300619a0a4003920aafe960e45688102",
                "requestBody": {
                    "description": "Данные нового комментария",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body",
                                    "user_id",
                                    "post_id"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Новый комментарий"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "post_id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комментарий успешно создан"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/comments/{id}": {
            "get": {
                "tags": [
                    "Комментарии"
                ],
                "summary": "Получить комментарий по ID",
                "operationId": "ca005bcffa22ab29b155dd7aa13b5d6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о комментарии"
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Комментарии"
                ],
                "summary": "Обновить комментарий",
                "operationId": "39e62963cbf897a1b2168683878e0a46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления комментария",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Обновленный комментарий"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "post_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комментарий успешно обновлён"
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Комментарии"
                ],
                "summary": "Удалить комментарий",
                "operationId": "721e5dc0c32ef81276203df965767606",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комментария для удаления",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Комментарий успешно удалён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комментарий не найден"
                    }
                }
            }
        },
...
    },
...
}